# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI-native headless workflow automation platform. DAG execution, MCP-native, queue-based.",
        "title": "aflow API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/webhooks/{workflow_id}": {
            "post": {
                "description": "Triggers a workflow execution via webhook. Authenticate with the secret from the publish response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Webhook trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow UUID",
                        "name": "workflow_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook secret (alternative to X-Webhook-Secret header)",
                        "name": "secret",
                        "in": "query"
                    },
                    {
                        "description": "Optional JSON payload (becomes execution input)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.webhookTriggerResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ProblemDetail"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api_handlers.ProblemDetail": {
            "description": "RFC 7807 error envelope returned on all 4xx/5xx responses.",
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Workflow abc was not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "title": {
                    "type": "string",
                    "example": "Not Found"
                },
                "type": {
                    "type": "string",
                    "example": "https://aflow.dev/errors/not-found"
                }
            }
        },
...
        "internal_api_handlers.webhookTriggerResp": {
            "type": "object",
            "properties": {
                "execution_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440010"
                },
                "status": {
                    "type": "string",
                    "example": "queued"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT (HS256, claim: workspace_id) or aflow_ API key. Set AFLOW_AUTH_DISABLED=true to skip in dev.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}